% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiddleR-package.R
\docType{package}
\name{fiddleR-package}
\alias{fiddleR}
\alias{fiddleR-package}
\title{fiddleR: A R wrapper for FIDDLE}
\description{
This package provides an R wrapper for the FIDDLE prepreprocessing pipeline.
}
\author{
\strong{Maintainer}: Michael Throolin \email{michael.throolin@utah.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
