% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_npz}
\alias{load_npz}
\title{Load a Sparse Matrix from an NPZ File}
\usage{
load_npz(npz_path, python_env = "test_env")
}
\arguments{
\item{npz_path}{A string specifying the path to the \code{.npz} file.}

\item{python_env}{A string specifying the name of the Python virtual environment to use. Defaults to \code{"test_env"}.}
}
\value{
A matrix or tensor loaded from the \code{.npz} file.
}
\description{
This function loads a sparse matrix from a \code{.npz} file using Python's \code{numpy} and \code{scipy.sparse} modules.
}
\examples{
\dontrun{
  s_matrix <- load_npz("path/to/file.npz", python_env = "test_env")
}
}
