% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fiddle.R
\name{run_fiddle}
\alias{run_fiddle}
\title{Run FIDDLE Python Script with R Data}
\usage{
run_fiddle(
  data_file,
  population_file,
  config_file,
  output_dir,
  T_ = 4,
  dt = 1,
  theta_1 = 0.001,
  theta_2 = 0.001,
  theta_freq = 1,
  stats_functions = c("min", "max", "mean"),
  impute_method = "ffill",
  python_env = "test_env"
)
}
\arguments{
\item{data_file}{A data frame or matrix to be used as the data input.}

\item{population_file}{Path to a vector or data frame representing the population.}

\item{config_file}{Path to config file.}

\item{output_dir}{Path to output directory.}

\item{T_}{Time period.}

\item{dt}{Time delta.}

\item{theta_1}{First theta value. Default is 0.001.}

\item{theta_2}{Second theta value. Default is 0.001.}

\item{theta_freq}{Theta frequency. Default is 1.0.}

\item{stats_functions}{Statistical functions to apply. Default is c("min", "max", "mean").}

\item{impute_method}{Method to use for imputing missing values ("ffill","mean","median", "linear", or "none"). Default is "ffill".}

\item{python_env}{Python virtual environment or path to Python executable.}
}
\value{
Outputs from the FIDDLE Python module.
}
\description{
This function runs the FIDDLE Python module with data passed as an R object.
}
