% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_python_env.R
\name{setup_python_env}
\alias{setup_python_env}
\title{Set Up Python Environment for Tests}
\usage{
setup_python_env(
  envname = "test_env",
  requirements_path = system.file("python/requirements.txt", package =
    utils::packageName()),
  fiddle_dir = system.file("python/FIDDLE", package = utils::packageName())
)
}
\arguments{
\item{envname}{Name of the virtual environment to create/use.}

\item{requirements_path}{Path to the \code{requirements.txt} file in python package.}

\item{fiddle_dir}{Path to the FIDDLE module directory.}
}
\description{
This function sets up a Python virtual environment, installs required
dependencies listed in the \code{requirements.txt} file, and ensures the
environment is ready for testing.
}
\examples{
# Set up Python environment for testing
setup_python_env(envname = "test_env")
}
