% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fiddle.R
\name{run_fiddle}
\alias{run_fiddle}
\title{Run FIDDLE Python Script with R Data}
\usage{
run_fiddle(
  data,
  population,
  T_,
  dt,
  theta_1 = 0.001,
  theta_2 = 0.001,
  theta_freq = 1,
  stats_functions = c("min", "max", "mean"),
  binarize = TRUE,
  python_env = "test_env"
)
}
\arguments{
\item{data}{A data frame or matrix to be used as the data input.}

\item{population}{A vector or data frame representing the population.}

\item{T_}{Time period.}

\item{dt}{Time delta.}

\item{theta_1}{First theta value. Default is 0.001.}

\item{theta_2}{Second theta value. Default is 0.001.}

\item{theta_freq}{Theta frequency. Default is 1.0.}

\item{stats_functions}{Statistical functions to apply. Default is c("min", "max", "mean").}

\item{binarize}{Binarize flag. Default is TRUE.}

\item{python_env}{Python virtual environment or path to Python executable.}
}
\value{
Outputs from the FIDDLE Python module.
}
\description{
This function runs the FIDDLE Python module with data passed as an R object.
}
